(maint) Add .bundle to gitignore and prep for 0.4.3
When building the module for publishing, Jenkins will install necessary
gems to `puppetlabs-python_task_helper/.bundle/` and if that directory
is not ignored will build the module tarball with `.bundle/` in it. This
adds `.bundle/` to the gitignore so it isn't included in module builds
(and isn't added to the git repo).